+++ alacritty/src/config/monitor.rs
@@ -1,4 +1,3 @@
-use std::fs;
 use std::path::PathBuf;
 use std::sync::mpsc;
 use std::time::Duration;
@@ -15,18 +14,34 @@ const DEBOUNCE_DELAY: Duration = Duration::from_millis(10);
 #[cfg(not(any(target_os = "linux", target_os = "macos", target_os = "windows")))]
 const DEBOUNCE_DELAY: Duration = Duration::from_millis(1000);
 
-pub fn watch(mut paths: Vec<PathBuf>, event_proxy: EventProxy) {
-    // Canonicalize all paths, filtering out the ones that do not exist.
-    paths = paths
-        .drain(..)
-        .filter_map(|path| match fs::canonicalize(&path) {
-            Ok(path) => Some(path),
-            Err(err) => {
-                error!("Unable to canonicalize config path {:?}: {}", path, err);
-                None
-            },
-        })
-        .collect();
+pub fn watch(paths: Vec<PathBuf>, event_proxy: EventProxy) {
+    let paths = {
+        let mut paths_tmp = Vec::new();
+        for path in paths {
+            if path.exists() {
+                paths_tmp.push(path.clone());
+
+                // Also watch the targets of symbolic links
+                match path.symlink_metadata() {
+                    Ok(metadata) if metadata.file_type().is_symlink() => {
+                        match path.canonicalize() {
+                            Ok(canonical_path) => paths_tmp.push(canonical_path),
+                            Err(err) => {
+                                error!("Unable to canonicalize config path {:?}: {}", path, err)
+                            },
+                        }
+                    },
+                    Ok(_) => {},
+                    Err(err) => {
+                        error!("Unable to get metadata for config path {:?}: {}", path, err)
+                    },
+                }
+            } else {
+                error!("Config path does not exist: {:?}", path);
+            }
+        }
+        paths_tmp
+    };
 
     // Don't monitor config if there is no path to watch.
     if paths.is_empty() {
@@ -73,17 +88,15 @@ pub fn watch(mut paths: Vec<PathBuf>, event_proxy: EventProxy) {
             };
 
             match event {
-                DebouncedEvent::Rename(..) => continue,
-                DebouncedEvent::Write(path)
+                DebouncedEvent::Rename(_, path)
+                | DebouncedEvent::Write(path)
                 | DebouncedEvent::Create(path)
-                | DebouncedEvent::Chmod(path) => {
-                    if !paths.contains(&path) {
-                        continue;
-                    }
-
+                | DebouncedEvent::Chmod(path)
+                    if paths.contains(&path) =>
+                {
                     // Always reload the primary configuration file.
                     event_proxy.send_event(Event::ConfigReload(paths[0].clone()));
-                },
+                }
                 _ => {},
             }
         }